make types dir inside src folder & clean some misunderstood commas
