+++ .gitignore
@@ -91,6 +91,9 @@ out
 .nuxt
 dist
 
+
+# IDE generator
+.idea
 # Gatsby files
 .cache/
 # Comment in the public line in if your project uses Gatsby and not Next.js
