+++ src/@types/ApplyType.ts
@@ -1,0 +1,9 @@
+/**
+ * @type {ApplyType}
+ * @description The type of the apply request
+ *
+ * @param {string} coupons The coupons type
+ * @param {string} giftcards The giftcards type
+ * @param {string} creator-codes The creator codes type
+ */
+export type ApplyType = "coupons" | "giftcards" | "creator-codes";
