+++ src/@types/BaseItem.ts
@@ -1,0 +1,12 @@
+
+/**
+ * @type {BaseItem}
+ * @description The base item object for the package and category objects
+ *
+ * @param {number} id The ID of the base item
+ * @param {string} name The name of the base item
+ */
+export type BaseItem = {
+    id: number;
+    name: string;
+}
