+++ src/@types/CouponCode.ts
@@ -1,0 +1,10 @@
+
+/**
+ * @type {CouponCode}
+ * @description The coupon code object for the body of the request
+ *
+ * @param {string} coupon_code The coupon code to apply or remove
+ */
+export type CouponCode = {
+    coupon_code: string;
+}
