+++ src/@types/CreatorCode.ts
@@ -1,0 +1,10 @@
+
+/**
+ * @type {CreatorCode}
+ * @description The creator code object for the body of the request
+ *
+ * @param {string} creator_code The creator code to apply or remove
+ */
+export type CreatorCode = {
+    creator_code: string;
+}
