+++ src/@types/Data.ts
@@ -1,0 +1,9 @@
+/**
+ * @type {Data}
+ * @description The data returned from the Tebex Headless API
+ *
+ * @type {T} The type of data to be returned
+ */
+export type Data<T> = {
+    data: T;
+}
