+++ src/@types/GenericObject.ts
@@ -1,0 +1,5 @@
+/**
+ * @type {GenericObject}
+ * @description A generic object
+ */
+export type GenericObject = string | number | boolean | null | undefined;
