+++ src/@types/GiftCardCode.ts
@@ -1,0 +1,10 @@
+
+/**
+ * @type {GiftCardCode}
+ * @description The gift card code object for the body of the request
+ *
+ * @param {string} card_number The gift card code to apply or remove
+ */
+export type GiftCardCode = {
+    card_number: string;
+}
