+++ src/@types/Route.ts
@@ -1,0 +1,8 @@
+/**
+ * @type {Route}
+ * @description The route of the Tebex Headless API
+ *
+ * @param {string} accounts The accounts route
+ * @param {string} baskets The baskets route
+ */
+export type Route = "accounts" | "baskets";
