+++ tsconfig.json
@@ -7,6 +7,11 @@
         "strict": true,
         "skipLibCheck": true,
         "noUncheckedIndexedAccess": true,
-        "noEmit": true
+        "noEmit": true,
+        "paths": {
+            "@custom-types/*": [
+                "./src/@types/*"
+            ]
+        }
     }
 }
